@@ -164,9 +164,9 @@ class Contents extends StatelessWidget {
                   onPressed: () async {
                     // Use the url_launcher plugin to open the Flutter docs in
                     // a browser.
-                    const url = 'https://flutter.dev/docs';
-                    if (await launcher.canLaunch(url)) {
-                      launcher.launch(url);
+                    final url = Uri.parse('https://flutter.dev/docs');
+                    if (await launcher.canLaunchUrl(url)) {
+                      await launcher.launchUrl(url);
                     }
                   },
                   child: const Text('Open Flutter Docs'),
